Moved weather widget to its own file in the integrations directory.
